requirements.txt
asgiref==3.3.1
da-vinci==0.3.0
Django==3.1.7
django-thumbnails==0.4.0
djangorestframework==3.12.4
Pillow==8.1.2
pytz==2021.1
shortuuid==1.0.1
sqlparse==0.4.1
